refactor: rename useKeyPairContext to Provider
