+++ src/components/[guild]/hooks/useAutoStatusUpdate.ts
@@ -1,5 +1,5 @@
 import { useWeb3React } from "@web3-react/core"
-import useKeyPair from "components/_app/useKeyPairContext"
+import { useKeyPair } from "components/_app/KeyPairProvider"
 import useMemberships from "components/explorer/hooks/useMemberships"
 import { mutateOptionalAuthSWRKey } from "hooks/useSWRWithOptionalAuth"
 import { useEffect } from "react"
