+++ src/components/_app/Web3ConnectionManager/components/WalletSelectorModal/components/DelegateCashButton.tsx
@@ -11,8 +11,8 @@ import {
   PopoverTrigger,
 } from "@chakra-ui/react"
 import { useWeb3React } from "@web3-react/core"
+import { useKeyPair } from "components/_app/KeyPairProvider"
 import { useWeb3ConnectionManager } from "components/_app/Web3ConnectionManager/Web3ConnectionManager"
-import useKeyPair from "components/_app/useKeyPairContext"
 import Button from "components/common/Button"
 import { ArrowSquareOut } from "phosphor-react"
