+++ src/components/_app/Web3ConnectionManager/hooks/useConnectFromLocalStorage.ts
@@ -1,8 +1,8 @@
-import useUser from "components/[guild]/hooks/useUser"
 import { useConnect } from "components/[guild]/JoinModal/hooks/useConnectPlatform"
 import { Message } from "components/[guild]/JoinModal/hooks/useOauthPopupWindow"
+import useUser from "components/[guild]/hooks/useUser"
+import { useKeyPair } from "components/_app/KeyPairProvider"
 import { usePostHogContext } from "components/_app/PostHogProvider"
-import useKeyPair from "components/_app/useKeyPairContext"
 import useToast from "hooks/useToast"
 import platforms from "platforms/platforms"
 import { useEffect } from "react"
