+++ src/components/common/Layout/components/Account/components/AccountModal/components/LinkDelegateVaultButton.tsx
@@ -1,4 +1,4 @@
-import useKeyPair from "components/_app/useKeyPairContext"
+import { useKeyPair } from "components/_app/KeyPairProvider"
 import Button from "components/common/Button"
 import Image from "next/image"
