+++ src/hooks/useSubmit/useSubmit.ts
@@ -3,7 +3,7 @@ import { keccak256 } from "@ethersproject/keccak256"
 import { JsonRpcProvider, Web3Provider } from "@ethersproject/providers"
 import { toUtf8Bytes } from "@ethersproject/strings"
 import { useWeb3React } from "@web3-react/core"
-import useKeyPair from "components/_app/useKeyPairContext"
+import { useKeyPair } from "components/_app/KeyPairProvider"
 import { Chains, RPC } from "connectors"
 import { randomBytes } from "crypto"
 import useLocalStorage from "hooks/useLocalStorage"
